cloud_function/requirements.txt
requests
pandas
beautifulsoup4
google-cloud-bigquery
python-dotenv
anthropic==0.64.0  
pyarrow
google-cloud-secret-manager
